Add option to change regtest chain config in CLI wallet
