+++ common/src/chain/config/mod.rs
@@ -17,6 +17,7 @@ mod builder;
 mod checkpoints;
 pub mod emission_schedule;
 pub mod regtest;
+pub mod regtest_options;
 pub use builder::Builder;
 use crypto::key::PublicKey;
 use crypto::vrf::VRFPublicKey;
