+++ node-lib/src/lib.rs
@@ -19,7 +19,6 @@ mod config_files;
 mod mock_time;
 pub mod node_controller;
 mod options;
-pub mod regtest_options;
 pub mod rpc;
 mod runner;
